naked url is the launcher, which will take you to wizard start or colinks app.  auth fixes with g-dawg
